build: add enqueueSnackbar and refactor some code
