+++ src/pages/index.ts
@@ -6,7 +6,7 @@ import Therapists from './therapists';
 import TherapistPage from './therapistProfile';
 import LoginAdmin from './loginAdmin';
 import NotFound from './notFound';
-import { ModelStyle } from './therapistProfile/classes';
+import { ModelStyle } from '../components/therapistProfile/therapistHeader/classes';
 
 export {
   Admin, LandingPage, Login, Signup, Therapists, TherapistPage, LoginAdmin, NotFound, ModelStyle,
